[GPU-Plugin] Fix gpu_hist to allow matrices with more than just 2^{32} elements limited by int (32-bit) and instead use bst_ulong or size_t for element-related quantities.  Tested with test_large.py .
Still need
* For larger matrix elements: src/common/hist_util.h std::vector<unsigned> row_ptr; -> std::vector<size_t> row_ptr;
* For larger max_bins > 1024: Currently fails to obtain correct train accuracy.

+++ plugin/updater_gpu/src/device_helpers.cuh
@@ -167,21 +167,21 @@ class range {
       return i_ < other.i_;
     }
 
-    __host__ __device__ void step(int s) { step_ = s; }
+    __host__ __device__ void step(int64_t s) { step_ = s; }
 
    protected:
     __host__ __device__ explicit iterator(int64_t start) : i_(start) {}
 
    public:
     uint64_t i_;
-    int step_ = 1;
+    int64_t step_ = 1;
   };
 
   __host__ __device__ iterator begin() const { return begin_; }
   __host__ __device__ iterator end() const { return end_; }
   __host__ __device__ range(int64_t begin, int64_t end)
       : begin_(begin), end_(end) {}
-  __host__ __device__ void step(int s) { begin_.step(s); }
+  __host__ __device__ void step(int64_t s) { begin_.step(s); }
 
  private:
   iterator begin_;
@@ -438,7 +438,7 @@ class bulk_allocator {
 
  public:
   ~bulk_allocator() {
-    for (int i = 0; i < d_ptr.size(); i++) {
+    for (size_t i = 0; i < d_ptr.size(); i++) {
       if (!(d_ptr[i] == nullptr)) {
         safe_cuda(cudaSetDevice(_device_idx[i]));
         safe_cuda(cudaFree(d_ptr[i]));
@@ -519,7 +519,7 @@ inline size_t available_memory(int device_idx) {
 template <typename T>
 void print(const thrust::device_vector<T> &v, size_t max_items = 10) {
   thrust::host_vector<T> h = v;
-  for (int i = 0; i < std::min(max_items, h.size()); i++) {
+  for (size_t i = 0; i < std::min(max_items, h.size()); i++) {
     std::cout << " " << h[i];
   }
   std::cout << "\n";
@@ -528,7 +528,7 @@ void print(const thrust::device_vector<T> &v, size_t max_items = 10) {
 template <typename T>
 void print(const dvec<T> &v, size_t max_items = 10) {
   std::vector<T> h = v.as_vector();
-  for (int i = 0; i < std::min(max_items, h.size()); i++) {
+  for (size_t i = 0; i < std::min(max_items, h.size()); i++) {
     std::cout << " " << h[i];
   }
   std::cout << "\n";
@@ -536,10 +536,10 @@ void print(const dvec<T> &v, size_t max_items = 10) {
 
 template <typename T>
 void print(char *label, const thrust::device_vector<T> &v,
-           const char *format = "%d ", int max = 10) {
+           const char *format = "%d ", size_t max = 10) {
   thrust::host_vector<T> h_v = v;
   std::cout << label << ":\n";
-  for (int i = 0; i < std::min(static_cast<int>(h_v.size()), max); i++) {
+  for (size_t i = 0; i < std::min(static_cast<size_t>(h_v.size()), max); i++) {
     printf(format, h_v[i]);
   }
   std::cout << "\n";
@@ -590,7 +590,7 @@ __global__ void launch_n_kernel(int device_idx, size_t begin, size_t end,
 template <int ITEMS_PER_THREAD = 8, int BLOCK_THREADS = 256, typename L>
 inline void launch_n(int device_idx, size_t n, L lambda) {
   safe_cuda(cudaSetDevice(device_idx));
-  const int GRID_SIZE = div_round_up(n, ITEMS_PER_THREAD * BLOCK_THREADS);
+  const int GRID_SIZE = div_round_up(n, ITEMS_PER_THREAD * BLOCK_THREADS); // FIXME: If n is large enough, this limits size can do.  Need to use 2D or 3D grid?
 #if defined(__CUDACC__)
   launch_n_kernel<<<GRID_SIZE, BLOCK_THREADS>>>(static_cast<size_t>(0), n,
                                                 lambda);
@@ -604,7 +604,7 @@ inline void multi_launch_n(size_t n, int n_devices, L lambda) {
   CHECK_LE(n_devices, n_visible_devices()) << "Number of devices requested "
                                               "needs to be less than equal to "
                                               "number of visible devices.";
-  const int GRID_SIZE = div_round_up(n, ITEMS_PER_THREAD * BLOCK_THREADS);
+  const int GRID_SIZE = div_round_up(n, ITEMS_PER_THREAD * BLOCK_THREADS); // FIXME: If n is large enough, this limits size can do.  Need to use 2D or 3D grid?
 #if defined(__CUDACC__)
   n_devices = n_devices > n ? n : n_devices;
   for (int device_idx = 0; device_idx < n_devices; device_idx++) {
@@ -653,11 +653,11 @@ struct BernoulliRng {
 template <typename func_t>
 class LauncherItr {
  public:
-  int idx;
+  size_t idx; // maybe?
   func_t f;
   XGBOOST_DEVICE LauncherItr() : idx(0) {}
-  XGBOOST_DEVICE LauncherItr(int idx, func_t f) : idx(idx), f(f) {}
-  XGBOOST_DEVICE LauncherItr &operator=(int output) {
+  XGBOOST_DEVICE LauncherItr(size_t idx, func_t f) : idx(idx), f(f) {}
+  XGBOOST_DEVICE LauncherItr &operator=(size_t output) {
     f(idx, output);
     return *this;
   }
@@ -700,7 +700,7 @@ class DiscardLambdaItr {
   XGBOOST_DEVICE DiscardLambdaItr(difference_type offset, func_t f)
       : offset(offset), f(f) {}
 
-  XGBOOST_DEVICE self_type operator+(const int &b) const {
+  XGBOOST_DEVICE self_type operator+(const size_t &b) const {
     return DiscardLambdaItr(offset + b, f);
   }
   XGBOOST_DEVICE self_type operator++() {
@@ -712,7 +712,7 @@ class DiscardLambdaItr {
     offset++;
     return retval;
   }
-  XGBOOST_DEVICE self_type &operator+=(const int &b) {
+  XGBOOST_DEVICE self_type &operator+=(const size_t &b) {
     offset += b;
     return *this;
   }
@@ -720,7 +720,7 @@ class DiscardLambdaItr {
     return LauncherItr<func_t>(offset, f);
   }
 
-  XGBOOST_DEVICE reference operator[](int idx) {
+  XGBOOST_DEVICE reference operator[](size_t idx) {
     self_type offset = (*this) + idx;
     return *offset;
   }
@@ -746,13 +746,13 @@ class DiscardLambdaItr {
  */
 
 template <typename func_t, typename segments_t>
-void TransformLbs(int device_idx, dh::CubMemory *temp_memory, int count,
+void TransformLbs(int device_idx, dh::CubMemory *temp_memory, size_t count,
                   thrust::device_ptr<segments_t> segments, int num_segments,
                   func_t f) {
   safe_cuda(cudaSetDevice(device_idx));
   auto counting = thrust::make_counting_iterator(0);
 
-  auto f_wrapper = [=] __device__(int idx, int upper_bound) {
+  auto f_wrapper = [=] __device__(size_t idx, int upper_bound) {
     f(idx, upper_bound - 1);
   };
